# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This API handles airline tickets.",
        "title": "MELI Bootcamp API",
        "termsOfService": "https://fake.url/",
        "contact": {
            "name": "API Support",
            "url": "https://real.url"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "localhost/8080",
    "paths": {
        "/ticket/getAverage/:dest": {
            "get": {
                "description": "Gets the percentage of total tickets bought over the last day with destination of given country",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Tickets"
                ],
                "summary": "Destination percentage",
                "parameters": [
                    {
                        "type": "string",
                        "description": "token",
                        "name": "token",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/ticket/getByCountry/:dest": {
            "get": {
                "description": "Gets the number of tickets bought for a trip to a given country passed by URL parameter",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Tickets"
                ],
                "summary": "Get tickets by country",
                "parameters": [
                    {
                        "type": "string",
                        "description": "token",
                        "name": "token",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        }
    }
}